Finalize CRUD on front end
Able to complete all CRUD actions
Had issue where was unable to update or destroy a post, but realized
that it is openRead, and though I could see the post, the user I was
signed in as wasn't the user associated with that post, so did not have
permission to edit/delete post.

Need to make edit/delete buttons available only on posts that are
written by current logged in user

Need to display user name as author of post
Need to compartmentalize code into components/templates now that all is
fully functional.